pyproject.toml
[project]
name = "cars-de-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mobile-de-parsing",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
]

[tool.uv.sources]
mobile-de-parsing = { git = "https://github.com/drogan-v/mobile_de.git" }
